Split up Justified collection specs into consistently named tests
